A6

+++ src/main/java/edu/iu/habahram/coffeeorder/model/Beverage.java
@@ -2,7 +2,7 @@ package edu.iu.habahram.coffeeorder.model;
 
 public abstract class Beverage {
 
-    public abstract float cost();
+    public abstract double cost();
 
     public abstract String getDescription();
