+++ src/main/java/edu/iu/habahram/coffeeorder/model/Decaf.java
@@ -2,7 +2,7 @@ package edu.iu.habahram.coffeeorder.model;
 
 public class Decaf extends Beverage{
     @Override
-    public float cost() { return 1.28F; }
+    public double cost() { return 1.28; }
 
     @Override
     public String getDescription() { return "Decaf"; }
