+++ src/main/java/edu/iu/habahram/coffeeorder/model/Espresso.java
@@ -2,7 +2,7 @@ package edu.iu.habahram.coffeeorder.model;
 
 public class Espresso extends Beverage{
     @Override
-    public float cost() { return 1.34F; }
+    public double cost() { return 1.34; }
 
     @Override
     public String getDescription() {
