+++ src/main/java/edu/iu/habahram/coffeeorder/model/HouseBlend.java
@@ -2,7 +2,7 @@ package edu.iu.habahram.coffeeorder.model;
 
 public class HouseBlend extends Beverage{
     @Override
-    public float cost() { return 1.65F; }
+    public double cost() { return 1.65; }
 
     @Override
     public String getDescription() {
