+++ src/main/java/edu/iu/habahram/coffeeorder/model/Receipt.java
@@ -1,6 +1,6 @@
 package edu.iu.habahram.coffeeorder.model;
 
-public record Receipt(String description, float cost) {
+public record Receipt(String description, double cost) {
     public String toStringWithId(int id) {
         return id + ", " + cost + ", " + description;
     }
